# docs/documentai-api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Document AI API",
    "description": "API for document processing",
    "version": "v1"
  },
  "paths": {
...
    "/v1/admin/audit-log/actions": {
      "get": {
        "tags": [
          "Admin:Audit Log"
        ],
        "summary": "Get Audit Actions",
        "description": "Return all known audit action strings.",
        "operationId": "get_audit_actions_v1_admin_audit_log_actions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditActionsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/v1/admin/extraction-eval": {
      "post": {
        "tags": [
          "Admin:Eval"
        ],
        "summary": "Run Eval",
        "description": "Submit a document through the real pipeline with both BDA and LLM extraction.\n\nReturns immediately with a job_id. Poll GET /{job_id} until 200.\n\nUses is_eval=True which:\n- Forces LLM extraction regardless of the feature flag\n- Writes each path's output to evalV1Responses map instead of terminal status\n- Suppresses metrics queue emission and batch counter increments",
        "operationId": "run_eval_v1_admin_extraction_eval_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_run_eval_v1_admin_extraction_eval_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvalSubmitResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/admin/extraction-eval/{job_id}": {
      "get": {
        "tags": [
          "Admin:Eval"
        ],
        "summary": "Get Eval Result",
        "description": "Poll for eval results. Returns 200 with results when both paths complete, 404 until then.",
        "operationId": "get_eval_result_v1_admin_extraction_eval__job_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvalResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuditActionsResponse": {
        "properties": {
          "actions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Actions"
          }
        },
        "type": "object",
        "required": [
          "actions"
        ],
        "title": "AuditActionsResponse"
      },
...
      "Body_run_eval_v1_admin_extraction_eval_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_run_eval_v1_admin_extraction_eval_post"
      },
...
      "EvalFieldResult": {
        "properties": {
          "value": {
            "title": "Value"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          }
        },
        "type": "object",
        "title": "EvalFieldResult"
      },
      "EvalResponse": {
        "properties": {
          "jobId": {
            "type": "string",
            "title": "Jobid"
          },
          "bda": {
            "additionalProperties": {
              "$ref": "#/components/schemas/EvalFieldResult"
            },
            "type": "object",
            "title": "Bda"
          },
          "llm": {
            "additionalProperties": {
              "$ref": "#/components/schemas/EvalFieldResult"
            },
            "type": "object",
            "title": "Llm"
          }
        },
        "type": "object",
        "required": [
          "jobId"
        ],
        "title": "EvalResponse"
      },
      "EvalSubmitResponse": {
        "properties": {
          "jobId": {
            "type": "string",
            "title": "Jobid"
          }
        },
        "type": "object",
        "required": [
          "jobId"
        ],
        "title": "EvalSubmitResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "API-Key"
      },
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}